third_party/WebKit/LayoutTests/http/tests/misc/performance-memory-in-dedicated-worker-expected.txt

Description

This test validates that performance.memory exists in dedicated workers


PASS self.performance is defined 
PASS self.performance.memory is defined 
PASS jsHeapSizeLimit should be no less than totalJSHeapSize 
PASS totalJSHeapSize should be no less than usedJSHeapSize 
